Files Location restructuring
