+++ config/webpack.config.babel.js
@@ -39,6 +39,7 @@ module.exports = {
           plugins: [
             "transform-decorators-legacy",
             "transform-class-properties",
+            "transform-export-extensions",
             [
               "transform-runtime", { "polyfill": false } // Fix for https://github.com/babel/babel/issues/2877
             ],
