+++ config/webpack.config.build.js
@@ -50,6 +50,7 @@ module.exports = {
           plugins: [
             "transform-decorators-legacy",
             "transform-class-properties",
+            "transform-export-extensions",
             [
               "transform-runtime", { "polyfill": false } // Fix for https://github.com/babel/babel/issues/2877
             ],
