+++ src/crudeditor-lib/check-model/index.js
@@ -1,0 +1,3 @@
+export checkModelDefinition from './modelDefinition.js';
+export checkSearchUi from './searchUi';
+export checkFormLayout from './formLayout';
