--- src/crudeditor-lib/lib/index.js
@@ -1,2 +1,0 @@
-export * from './lib';
-export * from './object-validation';
