+++ src/crudeditor-lib/views/create/reducer.js
@@ -1,7 +1,9 @@
 import cloneDeep from 'lodash/cloneDeep';
 import isEqual from 'lodash/isEqual';
 import u from 'updeep';
-import { checkFormLayout } from '../../lib';
+
+import { checkFormLayout } from '../../check-model';
+
 import {
   ALL_INSTANCE_FIELDS_VALIDATE,
